split plugin tests into separate files
